"MainScreen" - refactoring

+++ app/src/main/java/com/example/filmsappcompose/main_screen/domain/Film.kt
@@ -6,7 +6,7 @@ data class Film(
     val name: String,
     val image: String,
     val date_publication: Long,
-    val rating: Byte?,
+    val rating: Float?,
     val description: String,
     val age: String,
     val category: Int
